+++ OpenHD/ohd_interface/src/WBStreams.cpp
@@ -3,7 +3,7 @@
 
 #include "openhd-platform.hpp"
 #include "openhd-log.hpp"
-#include "openhd-wifi.hpp"
+#include "OHDWifiCard.hpp"
 #include "openhd-global-constants.h"
 
 #include <iostream>
